Pull out CMIS Atom feed fetcher
