+++ test/test_cmisatomfeed_fetcher.py
@@ -1,0 +1,59 @@
+# -*- coding: utf-8 -*-
+from unittest import TestCase
+import httpretty
+import harvester.fetcher as fetcher
+from test.utils import LogOverrideMixin
+from test.utils import DIR_FIXTURES
+
+
+class CMISAtomFeedFetcherTestCase(LogOverrideMixin, TestCase):
+    @httpretty.activate
+    def testCMISFetch(self):
+        httpretty.register_uri(
+            httpretty.GET,
+            'http://cmis-atom-endpoint/descendants',
+            body=open(DIR_FIXTURES+'/cmis-atom-descendants.xml').read())
+        h = fetcher.CMISAtomFeedFetcher(
+                'http://cmis-atom-endpoint/descendants',
+                'uname, pswd')
+        self.assertTrue(hasattr(h, 'objects'))
+        self.assertEqual(42, len(h.objects))
+
+    @httpretty.activate
+    def testFetching(self):
+        httpretty.register_uri(
+            httpretty.GET,
+            'http://cmis-atom-endpoint/descendants',
+            body=open(DIR_FIXTURES+'/cmis-atom-descendants.xml').read())
+        h = fetcher.CMISAtomFeedFetcher(
+                'http://cmis-atom-endpoint/descendants',
+                'uname, pswd')
+        num_fetched = 0
+        for obj in h:
+            num_fetched += 1
+        self.assertEqual(num_fetched, 42)
+
+
+# Copyright © 2016, Regents of the University of California
+# All rights reserved.
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+# - Redistributions of source code must retain the above copyright notice,
+#   this list of conditions and the following disclaimer.
+# - Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+# - Neither the name of the University of California nor the names of its
+#   contributors may be used to endorse or promote products derived from this
+#   software without specific prior written permission.
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
